@@ -1,6 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-  transpilePackages: ["lucide-react"],
-};
+const nextConfig = {};
 
-module.exports = nextConfig;
+export default nextConfig;
