@@ -1,19 +1,7 @@
 "use client";
 
 import * as React from "react";
-import {
-  AudioWaveform,
-  BookOpen,
-  Bot,
-  Command,
-  Frame,
-  GalleryVerticalEnd,
-  Map,
-  PieChart,
-  Plus,
-  Settings2,
-  SquareTerminal,
-} from "lucide-react";
+import { Plus } from "lucide-react";
 
 // import { NavMain } from "@/components/nav-main";
 import { NavHabbits } from "@/components/nav-habbits";
@@ -30,134 +18,134 @@ import { HabbitType } from "@/lib/types";
 import Link from "next/link";
 
 // This is sample data.
-const data = {
-  user: {
-    name: "shadcn",
-    email: "m@example.com",
-    avatar: "/avatars/shadcn.jpg",
-  },
-  // teams: [
-  //   {
-  //     name: "Acme Inc",
-  //     logo: GalleryVerticalEnd,
-  //     plan: "Enterprise",
-  //   },
-  //   {
-  //     name: "Acme Corp.",
-  //     logo: AudioWaveform,
-  //     plan: "Startup",
-  //   },
-  //   {
-  //     name: "Evil Corp.",
-  //     logo: Command,
-  //     plan: "Free",
-  //   },
-  // ],
-  // navMain: [
-  //   {
-  //     title: "Playground",
-  //     url: "#",
-  //     icon: SquareTerminal,
-  //     isActive: true,
-  //     items: [
-  //       {
-  //         title: "History",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Starred",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Settings",
-  //         url: "#",
-  //       },
-  //     ],
-  //   },
-  //   {
-  //     title: "Models",
-  //     url: "#",
-  //     icon: Bot,
-  //     items: [
-  //       {
-  //         title: "Genesis",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Explorer",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Quantum",
-  //         url: "#",
-  //       },
-  //     ],
-  //   },
-  //   {
-  //     title: "Documentation",
-  //     url: "#",
-  //     icon: BookOpen,
-  //     items: [
-  //       {
-  //         title: "Introduction",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Get Started",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Tutorials",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Changelog",
-  //         url: "#",
-  //       },
-  //     ],
-  //   },
-  //   {
-  //     title: "Settings",
-  //     url: "#",
-  //     icon: Settings2,
-  //     items: [
-  //       {
-  //         title: "General",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Team",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Billing",
-  //         url: "#",
-  //       },
-  //       {
-  //         title: "Limits",
-  //         url: "#",
-  //       },
-  //     ],
-  //   },
-  // ],
-  habbits: [
-    {
-      name: "Design Engineering",
-      url: "#",
-      icon: Frame,
-    },
-    {
-      name: "Sales & Marketing",
-      url: "#",
-      icon: PieChart,
-    },
-    {
-      name: "Travel",
-      url: "#",
-      icon: Map,
-    },
-  ],
-};
+// const data = {
+//   user: {
+//     name: "shadcn",
+//     email: "m@example.com",
+//     avatar: "/avatars/shadcn.jpg",
+//   },
+//   // teams: [
+//   //   {
+//   //     name: "Acme Inc",
+//   //     logo: GalleryVerticalEnd,
+//   //     plan: "Enterprise",
+//   //   },
+//   //   {
+//   //     name: "Acme Corp.",
+//   //     logo: AudioWaveform,
+//   //     plan: "Startup",
+//   //   },
+//   //   {
+//   //     name: "Evil Corp.",
+//   //     logo: Command,
+//   //     plan: "Free",
+//   //   },
+//   // ],
+//   // navMain: [
+//   //   {
+//   //     title: "Playground",
+//   //     url: "#",
+//   //     icon: SquareTerminal,
+//   //     isActive: true,
+//   //     items: [
+//   //       {
+//   //         title: "History",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Starred",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Settings",
+//   //         url: "#",
+//   //       },
+//   //     ],
+//   //   },
+//   //   {
+//   //     title: "Models",
+//   //     url: "#",
+//   //     icon: Bot,
+//   //     items: [
+//   //       {
+//   //         title: "Genesis",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Explorer",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Quantum",
+//   //         url: "#",
+//   //       },
+//   //     ],
+//   //   },
+//   //   {
+//   //     title: "Documentation",
+//   //     url: "#",
+//   //     icon: BookOpen,
+//   //     items: [
+//   //       {
+//   //         title: "Introduction",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Get Started",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Tutorials",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Changelog",
+//   //         url: "#",
+//   //       },
+//   //     ],
+//   //   },
+//   //   {
+//   //     title: "Settings",
+//   //     url: "#",
+//   //     icon: Settings2,
+//   //     items: [
+//   //       {
+//   //         title: "General",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Team",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Billing",
+//   //         url: "#",
+//   //       },
+//   //       {
+//   //         title: "Limits",
+//   //         url: "#",
+//   //       },
+//   //     ],
+//   //   },
+//   // ],
+//   habbits: [
+//     {
+//       name: "Design Engineering",
+//       url: "#",
+//       icon: Frame,
+//     },
+//     {
+//       name: "Sales & Marketing",
+//       url: "#",
+//       icon: PieChart,
+//     },
+//     {
+//       name: "Travel",
+//       url: "#",
+//       icon: Map,
+//     },
+//   ],
+// };
 
 export function AppSidebar({
   habbits,
@@ -168,7 +156,7 @@ export function AppSidebar({
   user: {
     name: string;
     email: string;
-    avatar: string;
+    avatar?: string;
   };
 }) {
   return (
